lava-stream-job: move config classes into lib/configuration.py
